requirements.txt
# Add Python dependencies here as the project evolves

## ingestion
yfinance
boto3
pandas

##Scheduling
apache-airflow==2.9.1

## analysis
#pandas
#ta
#scikit-learn